additional changes to make server more clear
