+++ backend/users/users.controller.js
@@ -5,6 +5,7 @@ const userService = require('./user.service');
 // routes
 router.post('/authenticate', authenticate);
 router.get('/', getAll);
+router.post('/registration', registration);
 
 module.exports = router;
 
@@ -18,4 +19,10 @@ function getAll(req, res, next) {
     userService.getAll()
         .then(users => res.json(users))
         .catch(err => next(err));
+}
+
+function registration(req, res){
+    userService.registration(req.body)
+        .then(data => data ? res.send(JSON.stringify({value: 'success'})) : res.send(JSON.stringify({value: 'fail'})))
+        .catch(err => console.log(err));
 }
